requirements.txt
google-generativeai
google-ai-gemini
